Add the location metadata as well as clean up the Nulecule file
This commit adds the 'location' metadata to each Nulecule file. Using
the FULL URL (due to different locations / docker container providers)

This is in-line to the future addition of the `atomicapp index` command
which will list supported and well documented Atomic App example
applications.

This commit also cleans up some changes to the Nulecule files in order
to make the formatting a bit better.